feat: consolidate checklist pages
